app/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
python-dotenv
pydantic
pydantic[email]
passlib[bcrypt]==1.7.4
bcrypt<4.0.0
python-jose
redis
greenlet
python-multipart
pyotp
cryptography[text]
qrcode
pillow